+++ lib/react_email_rails/render_modes.rb
@@ -1,1 +1,0 @@
-module ReactEmailRails::RenderModes; end